feat(matcher): validate URL Pattern at parse time, reject relative patterns
URL Pattern compilation errors used to be swallowed by Match, so a
subscriber sending matchURLPattern=/books/:id silently received nothing
with no diagnostic. The relative form has no base URL to resolve
against (topics are absolute IRIs and the protocol exposes no base) so
the spec now requires absolute patterns.

Introduce an optional PatternValidator interface; matchers that can
fail at compile time implement Validate. The hub calls it during
parseMatchers and resolveMatcherClaims, surfacing parse errors as 400
on the wire and 401 on JWT claims.

Drop the v8 compat workaround from the shutdown drain tests imported
during the rebase. They never tested deprecated behaviour, only the
shutdown contract, so move them to subscribe_shutdown_test.go without
the deprecated_topic build tag and switch the request to the v9 match=
parameter.
